Moving directories for clarification.
