+++ Plane/ServerScripts/plane.py
@@ -1,32 +1,34 @@
 import sys
-from flask import Flask, jsonify, request
-from flask_cors ipmort CORS
+import os
+
+script_dir = os.path.abspath('./../..')
+sys.path.append(script_dir)
 
 import General.Operations.arm as arm
 import General.Operations.initialize as initialize
 import General.Operations.mode as mode
+import General.Operations.speed as speed
+import General.Operations.waypoint_computation as waypoint_computation
 
+import Plane.Operations.systemState as systemState ## not fully configured, needs to be checked with a gps
+import Plane.Operations.takeoffConfiguration as takeoff_config
+import Plane.Operations.waypoint as waypoint
+import Plane.Operations.waypointConfiguration as waypoint_config
 
-app = Flask(__name__)
-CORS(app, resources={r"/*": {"origins" : "*"}}) # Overriding CORS for external access
+class Plane:
+    def __init__(self, vehicle_connection='udpin:127.0.0.1:14550'):
+        self.vehicle_connection = initialize.connect_to_vehicle(vehicle_connection)
+        print(f"Connnected to vehicle at {vehicle_connection}")
 
-vehicle_connection = None
+    def arm_vehicle(self):
+        arm.arm(self.vehicle_connection)
 
-if __name__ == '__main__':
-    if len(sus.argvv) == 2:
-        vehicle_port = sys.argv[1]
+    def set_flight_mode(self, mode_id):
+        mode.set_mode(self.vehicle_connection, mode_id)
 
-        BAUD = 57600
-        vehicle_port = 'COM6'
+    def set_speed(self, speed_input):
+        speed.set_cruise_speed(self.vehicle_connection, speed_input)
 
-        print(f"Attempting to connect to port: {vehicle_port}")
-        vehicle_connection = initialize.connect_to_vehicle(vehicle_port, BAUD)
-        print("Vehicle connection establish.")
-        retVal = initialize.verify_connection(vehicle_connection)
-        print("Vehicle connection verified.")
-
-        if retVal != True:
-            print("Connection failed. Exiting application.")
-            exit(1)
-
-        app.run(debug=True, host='0.0.0.0')
+    def add_waypoint(self, latitude, longitude, altitude):
+        waypoint.set_waypoint(self.vehicle_connection, latitude, longitude, altitude)
+
